+++ src/RuleFactory/RulesFactory/ConditionalRulesFactory.cs
@@ -1,5 +1,4 @@
 ﻿using RuleEngine.Rules;
-using RuleEngine.Rules.Conditional;
 
 namespace RuleFactory.RulesFactory
 {
